requirements.txt
diff-match-patch==20121119
Django==1.7.5
django-appconf==1.0.1
django-compressor==1.4
django-import-export==0.2.6
django-smtp-ssl==1.0
MySQL-python==1.2.5
Pillow==2.7.0
six==1.9.0
South==1.0.2
tablib==0.10.0
xlrd==0.9.3
xlutils==1.7.1
xlwt==1.0.0
